Fixing int -> long converstion issue

+++ src/main/java/org/broadinstitute/hellbender/tools/gvs/common/SchemaUtils.java
@@ -98,7 +98,7 @@ public class SchemaUtils {
     public static final long chromAdjustment = 1000000000000L;
 
     public static long encodeLocation(String chrom, int position) {
-        int chromosomeIndex = ChromosomeEnum.integerValueOfContig(chrom);
+        long chromosomeIndex = ChromosomeEnum.integerValueOfContig(chrom);
         return (long) chromosomeIndex * chromAdjustment + (long) position;
     }
 
@@ -110,7 +110,7 @@ public class SchemaUtils {
         Next 4 - state (16 values)
      */
     public static long encodeCompressedRefBlock(String chrom, long position, long length, long stateAsInt) {
-        int chromosomeIndex = ChromosomeEnum.integerValueOfContig(chrom);
+        long chromosomeIndex = ChromosomeEnum.integerValueOfContig(chrom);
         long packedBytes = ((0xFFFF & chromosomeIndex) << 48) |
                 ((0xFFFFFFFF & position) << 16) |
                 ((0xFFF & length) << 4) |
